Create Rank of a New Element
Your task is to program an algorithm that for an item and a sorted array, computes the rank of the item in the sorted array. More formally, for an item  and a sorted array , the rank of  is .

Please find a more detailed description of the task in the Main.java file. The task is to implement the method rank. If necessary, you chould create auxiliary methods. You do not need to change other methods.

You get one point for each passing test set. To to pass both test sets correctly, your solution has to be in O(log n) , where n  is the length of the input sorted array.