Create Length for the Longest Positive Subarray
Your task is to program an algorithm that for an array, computes the length of its longest positive subarray, where a subarray is called positive if all its elements are positive.

Please find a more detailed description of the task in the Main.java file. The task is to implement the method length_longest_positive_subarray. You are free to create auxiliary methods. You do not need to change other methods.

You get one point for each passing test set. To pass both test sets correctly, your solution has to be in O(n)  time, where n is the length of the input array.